+++ Dataset/files/meat_aliases.csv
@@ -1,0 +1,5 @@
+meat, alias1, alias2, ...
+beef, chuck, rib, ribs, brisket, plate, flank, loin, sirloin, round, shank
+pork, shoulder, cheek, belly, lard, ham, hock, tenderloin
+chicken
+lamb
