--- app/irsystem/ingredients/meat_aliases.csv
@@ -1,3 +1,0 @@
-meat, alias1, alias2, ...
-beef, chuck, rib, ribs, brisket, plate, flank, loin, sirloin, round, shank
-pork, shoulder, cheek, belly, lard, ham, hock, tenderloin
