+++ app/irsystem/ingredients_co2_src/README.txt
@@ -1,0 +1,7 @@
+Hello!
+
+If you are reading this, I made this directory because 
+I don't know where else to put things. This folder is probably
+temporary, and the files in it will probably be moved.
+
+-Liam
